Move FeedbackVotes permissions
